// src/gql/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
};
// ...
export type Breadcrumb = {
  __typename?: 'Breadcrumb';
  /**
   * The ID of the category.
   * @deprecated Use `category_uid` instead.
   */
  category_id: Maybe<Scalars['Int']['output']>;
  /** The category level. */
  category_level: Maybe<Scalars['Int']['output']>;
  /** The display name of the category. */
  category_name: Maybe<Scalars['String']['output']>;
  /** The unique ID for a `Breadcrumb` object. */
  category_uid: Scalars['ID']['output'];
  /** The URL key of the category. */
  category_url_key: Maybe<Scalars['String']['output']>;
  /** The URL path of the category. */
  category_url_path: Maybe<Scalars['String']['output']>;
};
// ...
export type CategoryInterface = {
  automatic_sorting: Maybe<Scalars['String']['output']>;
  available_sort_by: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  /** An array of breadcrumb items. */
  breadcrumbs: Maybe<Array<Maybe<Breadcrumb>>>;
  /** The relative canonical URL. This value is returned only if the system setting 'Use Canonical Link Meta Tag For Categories' is enabled. */
  canonical_url: Maybe<Scalars['String']['output']>;
  children_count: Maybe<Scalars['String']['output']>;
  /** Contains a category CMS block. */
  cms_block: Maybe<CmsBlock>;
  /**
   * The timestamp indicating when the category was created.
   * @deprecated The field should not be used on the storefront.
   */
  created_at: Maybe<Scalars['String']['output']>;
  custom_layout_update_file: Maybe<Scalars['String']['output']>;
  /** The attribute to use for sorting. */
  default_sort_by: Maybe<Scalars['String']['output']>;
  /** An optional description of the category. */
  description: Maybe<Scalars['String']['output']>;
  display_mode: Maybe<Scalars['String']['output']>;
  filter_price_range: Maybe<Scalars['Float']['output']>;
  /**
   * An ID that uniquely identifies the category.
   * @deprecated Use `uid` instead.
   */
  id: Maybe<Scalars['Int']['output']>;
  image: Maybe<Scalars['String']['output']>;
  include_in_menu: Maybe<Scalars['Int']['output']>;
  is_anchor: Maybe<Scalars['Int']['output']>;
  landing_page: Maybe<Scalars['Int']['output']>;
  /** The depth of the category within the tree. */
  level: Maybe<Scalars['Int']['output']>;
  meta_description: Maybe<Scalars['String']['output']>;
  meta_keywords: Maybe<Scalars['String']['output']>;
  meta_title: Maybe<Scalars['String']['output']>;
  /** The display name of the category. */
  name: Maybe<Scalars['String']['output']>;
  /** The full category path. */
  path: Maybe<Scalars['String']['output']>;
  /** The category path within the store. */
  path_in_store: Maybe<Scalars['String']['output']>;
  /** The position of the category relative to other categories at the same level in tree. */
  position: Maybe<Scalars['Int']['output']>;
  /** The number of products in the category that are marked as visible. By default, in complex products, parent products are visible, but their child products are not. */
  product_count: Maybe<Scalars['Int']['output']>;
  /** The list of products assigned to the category. */
  products: Maybe<CategoryProducts>;
  /** Indicates whether the category is staged for a future campaign. */
  staged: Scalars['Boolean']['output'];
  /** The unique ID for a `CategoryInterface` object. */
  uid: Scalars['ID']['output'];
  /**
   * The timestamp indicating when the category was updated.
   * @deprecated The field should not be used on the storefront.
   */
  updated_at: Maybe<Scalars['String']['output']>;
  /** The URL key assigned to the category. */
  url_key: Maybe<Scalars['String']['output']>;
  /** The URL path assigned to the category. */
  url_path: Maybe<Scalars['String']['output']>;
  /** The part of the category URL that is appended after the url key */
  url_suffix: Maybe<Scalars['String']['output']>;
};


/** Contains the full set of attributes that can be returned in a category search. */
export type CategoryInterfaceProductsArgs = {
  currentPage?: InputMaybe<Scalars['Int']['input']>;
  pageSize?: InputMaybe<Scalars['Int']['input']>;
  sort: InputMaybe<ProductAttributeSortInput>;
};

/** Contains details about the products assigned to a category. */
export type CategoryProducts = {
  __typename?: 'CategoryProducts';
  /** An array of products that are assigned to the category. */
  items: Maybe<Array<Maybe<ProductInterface>>>;
  /** Pagination metadata. */
  page_info: Maybe<SearchResultPageInfo>;
  /** The number of products in the category that are marked as visible. By default, in complex products, parent products are visible, but their child products are not. */
  total_count: Maybe<Scalars['Int']['output']>;
};

/** Contains a collection of `CategoryTree` objects and pagination information. */
export type CategoryResult = {
  __typename?: 'CategoryResult';
  /** A list of categories that match the filter criteria. */
  items: Maybe<Array<Maybe<CategoryTree>>>;
  /** An object that includes the `page_info` and `currentPage` values specified in the query. */
  page_info: Maybe<SearchResultPageInfo>;
  /** The total number of categories that match the criteria. */
  total_count: Maybe<Scalars['Int']['output']>;
};

/** Contains the hierarchy of categories. */
export type CategoryTree = CategoryInterface & RoutableInterface & {
  __typename?: 'CategoryTree';
  automatic_sorting: Maybe<Scalars['String']['output']>;
  available_sort_by: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  /** An array of breadcrumb items. */
  breadcrumbs: Maybe<Array<Maybe<Breadcrumb>>>;
  /** The relative canonical URL. This value is returned only if the system setting 'Use Canonical Link Meta Tag For Categories' is enabled. */
  canonical_url: Maybe<Scalars['String']['output']>;
  /** A tree of child categories. */
  children: Maybe<Array<Maybe<CategoryTree>>>;
  children_count: Maybe<Scalars['String']['output']>;
  /** Contains a category CMS block. */
  cms_block: Maybe<CmsBlock>;
  /**
   * The timestamp indicating when the category was created.
   * @deprecated The field should not be used on the storefront.
   */
  created_at: Maybe<Scalars['String']['output']>;
  custom_layout_update_file: Maybe<Scalars['String']['output']>;
  /** The attribute to use for sorting. */
  default_sort_by: Maybe<Scalars['String']['output']>;
  /** An optional description of the category. */
  description: Maybe<Scalars['String']['output']>;
  display_mode: Maybe<Scalars['String']['output']>;
  filter_price_range: Maybe<Scalars['Float']['output']>;
  /**
   * An ID that uniquely identifies the category.
   * @deprecated Use `uid` instead.
   */
  id: Maybe<Scalars['Int']['output']>;
  image: Maybe<Scalars['String']['output']>;
  include_in_menu: Maybe<Scalars['Int']['output']>;
  is_anchor: Maybe<Scalars['Int']['output']>;
  landing_page: Maybe<Scalars['Int']['output']>;
  /** The depth of the category within the tree. */
  level: Maybe<Scalars['Int']['output']>;
  meta_description: Maybe<Scalars['String']['output']>;
  meta_keywords: Maybe<Scalars['String']['output']>;
  meta_title: Maybe<Scalars['String']['output']>;
  /** The display name of the category. */
  name: Maybe<Scalars['String']['output']>;
  /** The full category path. */
  path: Maybe<Scalars['String']['output']>;
  /** The category path within the store. */
  path_in_store: Maybe<Scalars['String']['output']>;
  /** The position of the category relative to other categories at the same level in tree. */
  position: Maybe<Scalars['Int']['output']>;
  /** The number of products in the category that are marked as visible. By default, in complex products, parent products are visible, but their child products are not. */
  product_count: Maybe<Scalars['Int']['output']>;
  /** The list of products assigned to the category. */
  products: Maybe<CategoryProducts>;
  /** Contains 0 when there is no redirect error. A value of 301 indicates the URL of the requested resource has been changed permanently, while a value of 302 indicates a temporary redirect. */
  redirect_code: Scalars['Int']['output'];
  /** The internal relative URL. If the specified URL is a redirect, the query returns the redirected URL, not the original. */
  relative_url: Maybe<Scalars['String']['output']>;
  /** Indicates whether the category is staged for a future campaign. */
  staged: Scalars['Boolean']['output'];
  /** One of PRODUCT, CATEGORY, or CMS_PAGE. */
  type: Maybe<UrlRewriteEntityTypeEnum>;
  /** The unique ID for a `CategoryInterface` object. */
  uid: Scalars['ID']['output'];
  /**
   * The timestamp indicating when the category was updated.
   * @deprecated The field should not be used on the storefront.
   */
  updated_at: Maybe<Scalars['String']['output']>;
  /** The URL key assigned to the category. */
  url_key: Maybe<Scalars['String']['output']>;
  /** The URL path assigned to the category. */
  url_path: Maybe<Scalars['String']['output']>;
  /** The part of the category URL that is appended after the url key */
  url_suffix: Maybe<Scalars['String']['output']>;
};
// ...
export type CmsBlock = {
  __typename?: 'CmsBlock';
  /** The content of the CMS block in raw HTML. */
  content: Maybe<Scalars['String']['output']>;
  /** The CMS block identifier. */
  identifier: Maybe<Scalars['String']['output']>;
  /** The title assigned to the CMS block. */
  title: Maybe<Scalars['String']['output']>;
};
// ...
export type Query = {
// ...
  categories: Maybe<CategoryResult>;
// ...
};
// ...
export type RoutableInterface = {
  /** Contains 0 when there is no redirect error. A value of 301 indicates the URL of the requested resource has been changed permanently, while a value of 302 indicates a temporary redirect. */
  redirect_code: Scalars['Int']['output'];
  /** The internal relative URL. If the specified URL is a redirect, the query returns the redirected URL, not the original. */
  relative_url: Maybe<Scalars['String']['output']>;
  /** One of PRODUCT, CATEGORY, or CMS_PAGE. */
  type: Maybe<UrlRewriteEntityTypeEnum>;
};
// ...
export type SearchResultPageInfo = {
  __typename?: 'SearchResultPageInfo';
  /** The specific page to return. */
  current_page: Maybe<Scalars['Int']['output']>;
  /** The maximum number of items to return per page of results. */
  page_size: Maybe<Scalars['Int']['output']>;
  /** The total number of pages in the response. */
  total_pages: Maybe<Scalars['Int']['output']>;
};
// ...
export enum UrlRewriteEntityTypeEnum {
  Category = 'CATEGORY',
  CmsPage = 'CMS_PAGE',
  Product = 'PRODUCT',
  Pwa_404 = 'PWA_404'
}
// ...
export type GetNavigationDataQuery = { __typename?: 'Query', categories: { __typename?: 'CategoryResult', items: Array<{ __typename?: 'CategoryTree', children_count: string | null, children: Array<{ __typename?: 'CategoryTree', uid: string, level: number | null, name: string | null, url_path: string | null, url_key: string | null, position: number | null, children_count: string | null, children: Array<{ __typename?: 'CategoryTree', uid: string, level: number | null, name: string | null, url_path: string | null, url_key: string | null, position: number | null } | null> | null } | null> | null } | null> | null } | null };